backend/requirements.txt
Django==5.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
setuptools==71.1.0
psycopg2==2.9.9
eventlet
redis
pillow
python-dotenv
requests
django-phonenumber-field[phonenumbers]
drf-yasg==1.21.7